requirements.txt
Django==4.1
djangorestframework==3.14
markdown
django-filter 
environs==9.5
drf-yasg
djangorestframework-simplejwt
black
flake8